Change subtotal from integer to float and add subtotal to seed

+++ db/migrate/20170224194105_add_subtotal_to_order_donations.rb
@@ -1,5 +1,5 @@
 class AddSubtotalToOrderDonations < ActiveRecord::Migration[5.0]
   def change
-    add_column :order_donations, :subtotal, :integer
+    add_column :order_donations, :subtotal, :float
   end
 end
